requirements.txt
# Core dependencies
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1

# Data processing
pandas==2.1.4
numpy==1.24.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# API and Web
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0

# Integrations
python-telegram-bot==20.7
google-api-python-client==2.108.0
oauth2client==4.1.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
